fix height percentage for viewer
In the `<Viewer>` component, there was a problem with the calculation
of the pixels associated with the `y` coordinate and the `height`. This
is because the ratio (height / width) of the full image wasn't taken
into account, so these were being calculated under the assumption that
the image was square.

The fix is to calculate the ratio in the `<Navigator>` component and
pass that to the `<Viewer>`.